backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic
pydantic-settings
email-validator
ollama